Omit all fields in Image if empty.  Should only send when populated.
Without this the "Position" field was always set to 0, inserting the image at the beginning.  When omitted it is appended at the end.   You can set the Position explicitly to 1 to set it as the first image.  Same with other fields, they don't make sense to send on creation, or when updating.