Define types and constants for well-defined field values to aid in code editor type checking and auto-completion.
Shopify defines acceptable values for certain fields.  This defines those values, and an associated type, so that users are less likely to provide an unacceptable value.  Code editors may even provide a list of acceptable values for a field for auto-completion.  This is really just a code-writing improvement to prevent having to look up the acceptable values for a field in the Shopify API docs.

This should not break any existing code since the types are all strings and string values, in existing code, will still be accepted even though the fields' types aren't directly string.