Fix some linting warnings.
- Comment out unused code.
- Fix incorrect JSON struct tag (missing opening double-quote).
- Remove unneeded type inside slice of type ("redundant type from array, slice, or map composite literal" warning).

There are a lot of other warnings shown by staticcheck that were not fixed yet.  These all the warnings are for the checks listed below.  Fixing these is a rather big change.  Changing field names per ST1003 is a breaking change, so is probably best to be avoided.  Fixes for the other warnings is no issue since it is just a documentation change.

- ST1003 (using Api instead of API in struct field name),
- ST1021 (not starting the documentation of an exported type with it's name, and
- ST1020 (not starting the documentation of an exported method or func with the method's name).

Examples to be fixed...
abandoned_checkout.go:25:1: comment on exported type AbandonedCheckoutsResource should be of the form "AbandonedCheckoutsResource ..." (with optional leading article) (ST1021)
abandoned_checkout.go:85:1: comment on exported method List should be of the form "List ..." (ST1020)
blog.go:49:1: comment on exported type BlogResource should be of the form "BlogResource ..." (with optional leading article) (ST1021)
carrier.go:75:1: comment on exported type ShippingRateAddress should be of the form "ShippingRateAddress ..." (with optional leading article) (ST1021)
carrier.go:100:1: comment on exported type ShippingRateResponse should be of the form "ShippingRateResponse ..." (with optional leading article) (ST1021)
collect.go:39:1: comment on exported type CollectResource should be of the form "CollectResource ..." (with optional leading article) (ST1021)
collect.go:44:1: comment on exported type CollectsResource should be of the form "CollectsResource ..." (with optional leading article) (ST1021)
collection.go:39:1: comment on exported type CollectionResource should be of the form "CollectionResource ..." (with optional leading article) (ST1021)
collection.go:52:1: comment on exported method ListProducts should be of the form "ListProducts ..." (ST1020)